Hackery to use different serialize functions in isolated-vm vs in trusted code
runtime/common/marshaling/serializer.ts looks at an esbuild --define parameter
to figure out if it's being compiled for the isolated-vm thunk, and if so it
assumes that serialize functions have been set up in a global named
"codaInternal.serialize". Outside isolated-vm, there are simple functions to
wrap v8.serialize and deserialize, except encoded as a base64 string since
Buffer can't pass through the isolated-vm boundary.

There's also a new injectSerializer() function for bootstrapping the vm.

I haven't tested this on a demo yet, but the unit tests pass which seems like
a good sign.