Add "propertySchema" to PropertyAutocompleteFormulaSpecification
This seems like a simple and future-proof way to put the schema of the property currently being
autocompleted somewhere a pack maker can easily access it.

There are still some unresolved questions about whether propertyName and syncTable.schema should
or should be normalized, denormalized to the author-written values, or use the fromKey values
when they're set, but this may buy more time before we need to sort that out.