[2-way] Select list option to allow user-added strings
For string select lists we can support user-added strings if the pack maker allows it.

Objects are not supported because we'd need some way to ask the pack to construct and object from a string. In the future I think it could be nice to support some kind of a generic object builder so people can type in pack objects, but that's not in scope now.