call executeGetPermissions with actual sync from cli (#3162)
This PR adds the ability to get rowAccessDefinitions for a sync table from the CLI.

Previously, we had the ability to explicitly fetch permissions for rows, but you had to give the rows in a json parameter. This mode of Sync will instead do the sync fetching, pipe those into GetPermissions, and then return the result.